Create integration-pattern-between-powerapp-and-nav-system
Dynamics often requires connecting to on-premises Nav system. The purpose of this solution is to bring efficiency in processing data and information across the various operational departments within the organisation and to replace many of the disparate software solutions.
This Integration solution helps with seamless processing between Canvas App and on-premises Nav system.  Maintains data consistency between Dataverse and on premises Nav system by coordinating transactions across multiple services. This approach provides a centralized system for unique customer experience.